Refactor GitHub setup and add GPG key functionality
This commit refactors the GitHub setup functionality in the `vaultops.github_setup` module. It introduces the `setup_github` function to add vault access to GitHub user repositories and adds a GPG key to the bot GitHub account. Additionally, the `github_bot` module is added to provide functions for adding a GPG public key to a GitHub account using a Vault client.

The changes in this commit improve the organization and modularity of the codebase, making it easier to manage GitHub access and GPG key management.

Refactor release workflow to install poetry and update semantic-release dependencies

This commit refactors the release workflow to install Poetry and update the dependencies for semantic-release. This ensures that the release process is up-to-date and uses the latest versions of the required tools.

The changes in this commit improve the reliability and compatibility of the release workflow, ensuring that the project can be released smoothly with the latest tools and dependencies.

Update models related to PKI private keys and root tokens

This commit updates the models related to PKI private keys and root tokens in the `vaultops.models` module. It introduces the `PrivateKeyProperties` and `GeneratedPrivateKey` classes to represent the properties and generation status of private keys. Additionally, the `VaultNewRootToken` class is added to represent the response when generating a new root token in HashiCorp Vault.

These changes enhance the clarity and consistency of the codebase, providing more descriptive and structured models for PKI private keys and root tokens.

Update Vault configuration and secrets model

This commit updates the `VaultConfig` class in the `vaultops.models.vault_config` module. It improves the representation of the configuration and secrets required for interacting with HashiCorp Vault.

The changes in this commit enhance the readability and maintainability of the codebase, making it easier to work with Vault configuration and secrets.

Update models related to Vault Raft nodes

This commit updates the models related to Vault Raft nodes in the `vaultops.models` module. It introduces the `VaultRaftNode` and `VaultRaftNodeHvac` classes to represent Vault Raft nodes with HVAC client configuration.

These changes improve the flexibility and extensibility of the codebase, allowing for easier management and configuration of Vault Raft nodes.